requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.2.2
Django==1.11.13
django-pure-pagination==0.3.0
dnspython==1.15.0
enum34==1.1.6
idna==2.7
ipaddress==1.0.22
mysqlclient==1.3.12
Pillow==5.1.0
pycparser==2.18
pytz==2018.4
six==1.11.0
